+++ flv/flv_test.go
@@ -1,4 +1,4 @@
-package av
+package flv
 
 import (
 	"testing"
